crank4py-connector/requirements.txt
six
websocket-client==0.57.0
yarl
requests
